Rename site-ocpi to ocpi
